requirements.txt
langchain
langchain-community
langchain-pinecone
langchain-groq
langchain_openai
langchain_experimental
langchain-core
flask
pypdf
python-dotenv
pinecone[grpc]
sentence-transformers==2.2.2
-e .